+++ src/test/resources/ru/vyarus/gradle/plugin/quality/java/sample/Sample.java
@@ -1,5 +1,8 @@
 package sample;
 
+import java.util.ArrayDeque;
+import java.util.Deque;
+
 public class Sample {
 
     private String sample;
@@ -9,6 +12,7 @@ public class Sample {
     }
 
     public static void main(String[] args) {
-
+        final Deque res = new ArrayDeque();
+        System.out.println(res);
     }
 }